+++ erppeek.py
@@ -64,6 +64,7 @@ __version__ = '0.12.dev0'
 __all__ = ['Client', 'Model', 'Record', 'RecordList', 'Service', 'read_config']
 
 CONF_FILE = 'erppeek.ini'
+HIST_FILE = os.path.expanduser('~/.erppeek_history')
 DEFAULT_URL = 'http://localhost:8069'
 DEFAULT_DB = 'openerp'
 DEFAULT_USER = 'admin'
@@ -1196,9 +1197,19 @@ def _interact(use_pprint=True, usage=USAGE):
     builtins.usage = Usage()
 
     try:
-        __import__('readline')
+        import readline as rl
+        import rlcompleter
     except ImportError:
         pass
+    else:
+        import atexit
+        rl.parse_and_bind('tab: complete')
+        if os.path.exists(HIST_FILE):
+            rl.read_history_file(HIST_FILE)
+            if rl.get_history_length() < 0:
+                rl.set_history_length(int(os.environ.get('HISTSIZE', 500)))
+            # better append instead of replace?
+            atexit.register(rl.write_history_file, HIST_FILE)
 
     class Console(code.InteractiveConsole):
         def runcode(self, code):